requirements.txt
alembic==1.13.3
amarium==0.2.12
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
bcrypt==4.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.4
cryptography==43.0.3
dill==0.3.9
dnspython==2.7.0
docformatter==1.7.5
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.1.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prettify-py==0.2.3.12
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pylint==3.3.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.2
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
tomlkit==0.13.2
typer==0.12.5
typing_extensions==4.12.2
untokenize==0.1.1
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
